requirements.txt
httpx==0.28.1
python-dotenv==1.1.0
boto3==1.38.0
langchain==0.3.25
langchain-aws==0.2.22
langgraph==0.4.1
fastapi==0.115.12
uvicorn==0.34.2
psycopg2-binary==2.9.10
pgvector==0.4.1
beautifulsoup4==4.13.4
tiktoken==0.9.0
ragas==0.2.15
langsmith==0.3.42
